@@ -6,17 +6,16 @@
 
     <artifactId>leaflet-rest-clients-pack</artifactId>
     <packaging>pom</packaging>
-    <version>1.12.2</version>
+    <version>1.13.0</version>
 
     <parent>
         <groupId>hu.psprog.leaflet</groupId>
         <artifactId>leaflet-stack-base-bom</artifactId>
-        <version>2.3-r2303.2</version>
+        <version>3.0-r2305.1</version>
     </parent>
 
     <modules>
         <module>backend-rest-client</module>
-        <module>rcp-bundle</module>
         <module>tms-rest-client</module>
         <module>failover-rest-client</module>
         <module>tlp-rest-client</module>
@@ -26,7 +25,6 @@
         <module>recaptcha-rest-api</module>
         <module>recaptcha-rest-client</module>
         <module>request-adapters</module>
-        <module>rcp-hystrix-support</module>
         <module>lens-rest-api</module>
         <module>lens-rest-client</module>
         <module>lsrs-rest-api</module>
@@ -36,7 +34,7 @@
     <properties>
 
         <!-- Java environment version -->
-        <java.version>11</java.version>
+        <java.version>17</java.version>
 
         <!-- overrides for leaflet internal libs -->
         <leaflet.rcp.version>${project.version}</leaflet.rcp.version>
