@@ -1,7 +1,5 @@
 package hu.psprog.leaflet.bridge.service.impl;
 
-import com.netflix.hystrix.contrib.javanica.annotation.DefaultProperties;
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import hu.psprog.leaflet.api.rest.request.tag.TagAssignmentRequestModel;
 import hu.psprog.leaflet.api.rest.request.tag.TagCreateRequestModel;
 import hu.psprog.leaflet.api.rest.response.common.WrapperBodyDataModel;
@@ -10,13 +8,12 @@ import hu.psprog.leaflet.api.rest.response.tag.TagListDataModel;
 import hu.psprog.leaflet.bridge.client.BridgeClient;
 import hu.psprog.leaflet.bridge.client.domain.BridgeService;
 import hu.psprog.leaflet.bridge.client.exception.CommunicationFailureException;
-import hu.psprog.leaflet.bridge.config.LeafletPath;
 import hu.psprog.leaflet.bridge.client.request.RESTRequest;
 import hu.psprog.leaflet.bridge.client.request.RequestMethod;
+import hu.psprog.leaflet.bridge.config.LeafletPath;
 import hu.psprog.leaflet.bridge.service.TagBridgeService;
+import jakarta.ws.rs.core.GenericType;
 import org.springframework.beans.factory.annotation.Autowired;
-
-import javax.ws.rs.core.GenericType;
 
 /**
  * Implementation of {@link TagBridgeService}.
@@ -24,12 +21,11 @@ import javax.ws.rs.core.GenericType;
  * @author Peter Smith
  */
 @BridgeService(client = "leaflet")
-@DefaultProperties(groupKey = "leaflet.content")
-public class TagBridgeServiceImpl extends HystrixDefaultConfiguration implements TagBridgeService {
+public class TagBridgeServiceImpl implements TagBridgeService {
 
     private static final String ID = "id";
 
-    private BridgeClient bridgeClient;
+    private final BridgeClient bridgeClient;
 
     @Autowired
     public TagBridgeServiceImpl(BridgeClient bridgeClient) {
@@ -49,7 +45,6 @@ public class TagBridgeServiceImpl extends HystrixDefaultConfiguration implements
     }
 
     @Override
-    @HystrixCommand
     public WrapperBodyDataModel<TagListDataModel> getAllPublicTags() throws CommunicationFailureException {
 
         RESTRequest restRequest = RESTRequest.getBuilder()
