@@ -2,12 +2,15 @@ package hu.psprog.leaflet.bridge.service.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import hu.psprog.leaflet.api.rest.response.sitemap.Sitemap;
+import hu.psprog.leaflet.api.rest.response.sitemap.SitemapLocationItem;
 import hu.psprog.leaflet.bridge.client.exception.CommunicationFailureException;
 import hu.psprog.leaflet.bridge.config.LeafletPath;
 import hu.psprog.leaflet.bridge.it.config.BridgeITSuite;
 import hu.psprog.leaflet.bridge.service.SitemapBridgeService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+
+import java.util.List;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
 import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
@@ -26,9 +29,11 @@ import static org.hamcrest.MatcherAssert.assertThat;
 public class SitemapBridgeServiceImplIT extends WireMockBaseTest {
 
     private static final Sitemap SITEMAP = Sitemap.getBuilder()
-            .withLocation("/location-1")
-            .withLocation("/location-2")
-            .withLocation("/location/sub/3")
+            .withSitemapLocationItemList(List.of(
+                    new SitemapLocationItem("/location-1"),
+                    new SitemapLocationItem("/location-2"),
+                    new SitemapLocationItem("/location/sub/3")
+            ))
             .build();
 
     @Autowired
