@@ -1,7 +1,6 @@
 package hu.psprog.leaflet.lens.api.domain;
 
 import lombok.Builder;
-import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
 import java.util.Map;
@@ -11,20 +10,19 @@ import java.util.Map;
  *
  * @author Peter Smith
  */
-@Data
 @Builder
 @Jacksonized
-public class ContactRequest implements MailContent {
+public record ContactRequest(
+        String name,
+        String email,
+        String message
+) implements MailContent {
 
     public static final String MAIL_CONTENT_TYPE = "contact_request";
 
     private static final String NAME = "name";
     private static final String EMAIL = "email";
     private static final String MESSAGE = "message";
-
-    private final String name;
-    private final String email;
-    private final String message;
 
     @Override
     public Map<String, Object> asContentMap() {
