@@ -1,9 +1,7 @@
 package hu.psprog.leaflet.lsrs.api.response;
 
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import lombok.Builder;
-import lombok.Data;
+import lombok.extern.jackson.Jacksonized;
 
 import java.util.List;
 
@@ -12,15 +10,8 @@ import java.util.List;
  *
  * @author Peter Smith
  */
-@Data
 @Builder
-@JsonDeserialize(builder = FileListDataModel.FileListDataModelBuilder.class)
-public class FileListDataModel {
-
-    private List<FileDataModel> files;
-
-    @JsonPOJOBuilder(withPrefix = "")
-    public static class FileListDataModelBuilder {
-
-    }
-}
+@Jacksonized
+public record FileListDataModel(
+        List<FileDataModel> files
+) { }
