@@ -1,8 +1,6 @@
 package hu.psprog.leaflet.tlp.api.domain;
 
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
+import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.util.Date;
@@ -13,6 +11,7 @@ import java.util.Objects;
  *
  * @author Peter Smith
  */
+@Data
 public class LogRequest {
 
     private boolean queried;
@@ -30,55 +29,6 @@ public class LogRequest {
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
     private Date to;
 
-    public boolean isQueried() {
-
-        return queried;
-    }
-
-    public void setQueried(boolean queried) {
-        this.queried = queried;
-    }
-
-    public String getSource() {
-        return source;
-    }
-
-    public void setSource(String source) {
-        this.source = source;
-    }
-
-    public String getLevel() {
-        return level;
-    }
-
-    public void setLevel(String level) {
-        this.level = level;
-    }
-
-    public Date getFrom() {
-        return from;
-    }
-
-    public void setFrom(Date from) {
-        this.from = from;
-    }
-
-    public Date getTo() {
-        return to;
-    }
-
-    public void setTo(Date to) {
-        this.to = to;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
     public int getPage() {
 
         if (page == 0) {
@@ -86,10 +36,6 @@ public class LogRequest {
         }
 
         return page;
-    }
-
-    public void setPage(int page) {
-        this.page = page;
     }
 
     public int getLimit() {
@@ -101,10 +47,6 @@ public class LogRequest {
         return limit;
     }
 
-    public void setLimit(int limit) {
-        this.limit = limit;
-    }
-
     public OrderBy getOrderBy() {
 
         if (Objects.isNull(orderBy)) {
@@ -114,10 +56,6 @@ public class LogRequest {
         return orderBy;
     }
 
-    public void setOrderBy(OrderBy orderBy) {
-        this.orderBy = orderBy;
-    }
-
     public OrderDirection getOrderDirection() {
 
         if (Objects.isNull(orderDirection)) {
@@ -125,63 +63,5 @@ public class LogRequest {
         }
 
         return orderDirection;
-    }
-
-    public void setOrderDirection(OrderDirection orderDirection) {
-        this.orderDirection = orderDirection;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-
-        if (o == null || getClass() != o.getClass()) return false;
-
-        LogRequest that = (LogRequest) o;
-
-        return new EqualsBuilder()
-                .append(queried, that.queried)
-                .append(page, that.page)
-                .append(limit, that.limit)
-                .append(source, that.source)
-                .append(level, that.level)
-                .append(from, that.from)
-                .append(to, that.to)
-                .append(content, that.content)
-                .append(orderBy, that.orderBy)
-                .append(orderDirection, that.orderDirection)
-                .isEquals();
-    }
-
-    @Override
-    public int hashCode() {
-        return new HashCodeBuilder(17, 37)
-                .append(queried)
-                .append(source)
-                .append(level)
-                .append(from)
-                .append(to)
-                .append(content)
-                .append(page)
-                .append(limit)
-                .append(orderBy)
-                .append(orderDirection)
-                .toHashCode();
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .append("queried", queried)
-                .append("source", source)
-                .append("level", level)
-                .append("from", from)
-                .append("to", to)
-                .append("content", content)
-                .append("page", page)
-                .append("limit", limit)
-                .append("orderBy", orderBy)
-                .append("orderDirection", orderDirection)
-                .toString();
     }
 }
